#231 Move Stateless component to global and other to separate file
